+++ src/main/java/com/ugav/algo/RMQPlusMinusOneBenderFarachColton2000.java
@@ -73,8 +73,8 @@ public class RMQPlusMinusOneBenderFarachColton2000 extends RMQLinearAbstract {
 		for (int b = 0; b < ds.blockNum; b++) {
 			int key = calcBlockKey(ds, b);
 
-			ds.interBlocksDs[b] = tables.computeIfAbsent(key, k -> {
-				int demoBlock[] = calcDemoBlock(k, ds.blockSize);
+			ds.interBlocksDs[b] = tables.computeIfAbsent(Integer.valueOf(key), k -> {
+				int demoBlock[] = calcDemoBlock(k.intValue(), ds.blockSize);
 				return RMQLookupTable.getInstace().preprocessRMQ(new ArrayIntComparator(demoBlock), demoBlock.length);
 			});
 		}
@@ -91,7 +91,7 @@ public class RMQPlusMinusOneBenderFarachColton2000 extends RMQLinearAbstract {
 
 		@Override
 		int getBlockSize(int n) {
-			return n  <= 1 ? 1 : (int) Math.ceil(Utils.log2((double) n) / 2);
+			return n <= 1 ? 1 : (int) Math.ceil(Utils.log2((double) n) / 2);
 		}
 
 		@Override
